Refactor tooltip charts into separate files
