+++ packages/components/src/highlight-cards/charts/comparingInfoRange.tsx
@@ -1,0 +1,12 @@
+const comparingInfoRangeChart = (
+	<svg width="132" height="4" viewBox="0 0 132 4" fill="none" xmlns="http://www.w3.org/2000/svg">
+		<rect x="0.5" width="66" height="1" fill="#A7AAAD" />
+		<rect x="66.5" y="1" width="3" height="1" transform="rotate(90 66.5 1)" fill="#A7AAAD" />
+		<rect x="1.5" y="1" width="3" height="1" transform="rotate(90 1.5 1)" fill="#A7AAAD" />
+		<rect x="65.5" width="66" height="1" fill="#F6F7F7" />
+		<rect x="131.5" y="1" width="3" height="1" transform="rotate(90 131.5 1)" fill="#F6F7F7" />
+		<rect x="66.5" y="1" width="3" height="1" transform="rotate(90 66.5 1)" fill="#F6F7F7" />
+	</svg>
+);
+
+export default comparingInfoRangeChart;
