+++ packages/components/src/highlight-cards/charts/index.tsx
@@ -1,0 +1,2 @@
+export { default as comparingInfoBarsChart } from './comparingInfoBars';
+export { default as comparingInfoRangeChart } from './comparingInfoRange';
